+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Option to pick a template directory when not configured:
   - Error message now gives the option to pick a template directory using the systems dialog.
+- Folder in default structure:
+  - The default structure now contains 'images' with the template 'folder' as example on how to add folders.
 
 ## [1.2.1] – 2025-08-10
 ### Added